Fix github action itegration test execution
The github action was using the same configuration as the template
repository which tests both containers and machine varients of the
teampltes. This doesn't apply to to a charm as it will be created with
only one and the action therefore doesn't need to select.

This isn't covered in the integration testing because a github runner is
not setup to verify github action configuration.